Crate five8_core

five8_core

This crate contains core utilities for the five8 and five8_const crates.

It’s unlikely you’ll need anything from here except DecodeError, BASE58_ENCODED_32_MAX_LEN, and BASE58_ENCODED_64_MAX_LEN. For semver purposes, these are the only things that are part of the public API. The rest is for internal use only and will not follow semver.

Enums

DecodeError

Constants

BASE58_ENCODED_32_MAX_LEN

BASE58_ENCODED_64_MAX_LEN

BASE58_INVALID_CHAR

BASE58_INVERSE

BASE58_INVERSE_TABLE_OFFSET

BASE58_INVERSE_TABLE_SENTINEL

BINARY_SZ_32

BINARY_SZ_64

DEC_TABLE_32

DEC_TABLE_64

INTERMEDIATE_SZ_32

INTERMEDIATE_SZ_64

N_32

N_64

RAW58_SZ_32

RAW58_SZ_64